get-syslog.py: retrieve the system log
This downloads the syslog from the gigahub. This is exactly the same as
what you would see in the web interface under System Log.

Note that the name associated with each MAC address seems to be handled
dynamically. If you reboot the modem, it "forgets" the associated name,
and shows only the MAC address in the syslog. If it later gets a new
name for the MAC, this will be shown *everywhere* in the log. It's
almost like they only store the MAC, and do a "lookup" against its name
each time the syslog is requested.

Signed-off-by: Ralph Siemsen <ralphs@netwinder.org>